Refactor type casting and improve code clarity
- Updated various functions to use explicit type casting for size calculations and comparisons, enhancing code readability and preventing potential issues.
- Changed multiple instances of `char *` to `const char *` to enforce immutability where applicable, improving code safety.
- Added `[[maybe_unused]]` annotations to parameters in several functions to suppress compiler warnings for unused variables.
- Improved consistency in handling vector operations and function signatures across the codebase.

+++ mymod/gamesrc/cgame/cg_challenges.c
@@ -54,7 +54,7 @@ void challenges_init(void) {
 	fileLen = trap_FS_FOpenFile(FILENAME, &file, FS_READ);
 
 	//If the file not is big enough to contain all challenges then initialize from zero
-	if (fileLen<sizeof (challengeTable)) {
+	if (fileLen < (int)sizeof (challengeTable)) {
 		trap_FS_FCloseFile(file); //Lets remeber to close the file before all returns
 		for (i = 0; i < CHALLENGES_MAX; i++)
 			challengeTable[i] = 0; //Init all challenges to zero
